backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
google-genai
python-multipart
mysql-connector-python